+++ src/sbuf.h
@@ -95,7 +95,7 @@ struct __cex_namespace__sbuf {
     /// Append format va (using CEX formatting engine), always null-terminating
     Exc             (*appendfva)(sbuf_c* self, char* format, va_list va);
     /// Returns string capacity from its metadata
-    u32             (*capacity)(sbuf_c* self);
+    usize           (*capacity)(sbuf_c* self);
     /// Clears string
     void            (*clear)(sbuf_c* self);
     /// Creates new dynamic string builder backed by allocator
@@ -107,7 +107,7 @@ struct __cex_namespace__sbuf {
     /// Returns false if string invalid
     bool            (*isvalid)(sbuf_c* self);
     /// Returns string length from its metadata
-    u32             (*len)(sbuf_c* self);
+    usize           (*len)(sbuf_c* self);
     /// Sets the length of a string to any value, if new_length greater than capacity, re-allocates more
     /// space, always null-terminating. Newly allocated space is not ZII'ed, you must fill it yourself.
     Exc             (*set_len)(sbuf_c* self, usize new_length);
